A3: Add "I think it is" command + add correct asnwer counter function
I added a third command. The user can say the answer orally and it will be checked through the oralAnswer function to see if it is correct. If so it turns green. Otherwise, the button shakes. I also added a counter function which counts number of correct answers. If the answer is right, it is added to the total number.