A3: Moved css properties to css file + Added correct answer counter
I moved some of my elements css properties to css file to have a shorter and cleaner code. I also add correct answer counter. So, everytime the answer is correct the button turns green and one is added to the counter. If answer is wrong it shakes and gets removed. The counter is reset to zero as well.