P2: Added give up command + play again button + warning message
I added a give up command. Therefore, if the player couldn't answer the question, he/she is not stopped by that question and would be able to continue. I added a play again button so the player can play again. I also added a warning message which let them know if they are giving the right answer or not.