P2: Added sound for wild animals
I added sound for wild animals to make the game more scary when they appear on the screen. As the number of wild animals that I have is limited, I used switch statement instead of using a for loop and doing kind of the same thing that I did for images.