A4: Gave a story and meaning to the description
I made a separate function for the random elements which get from the data to have a nicer and cleaner code. I made a story for the description so it would be more interesting to interact with. And, I added an "andbutton" so the user knows if they click on it, they can see more random description.