P3: Finalized the start screen and level one screen to 70%
I am 70% done with my game screen and I only need to add the last puzzle + replacing hard coded numbers by variables and adding the reference. This time, I added the pop-up window which is shown once the puzzle is complete. It contains two button which one of them send the player to home page and ther other one reloads the page.